Rename BitcoinAddress to CkBTCInfoCard
