+++ frontend/src/lib/components/accounts/CkBTCInfoCard.svelte
